refactor: create exercises/ dir everywhere
