+++ .vscode/settings.json
@@ -37,7 +37,8 @@
 		"components/ui": "core",
 		"components/fragments": "core",
 		"geo-data/languages": "",
-		"[[...query]]": "api"
+		"[[...query]]": "api",
+		"exercises": "vscode"
 	},
 	// https://youtu.be/3_-Je5XpbqY?t=688
 	// https://stackoverflow.com/a/76125894/6543935
