+++ services/exercises/user-service.ts
@@ -1,0 +1,3 @@
+import create from "@/services/exercises/http-service";
+
+export default create("/users");
