--- services/user-service.ts
@@ -1,3 +1,0 @@
-import create from "@/services/http-service";
-
-export default create("/users");
